tests/test_requirements.txt
pre-commit
black
flake8
pytest
pytest-cov
mypy
freezegun
GitPython